+++ client/src/extension.ts
@@ -6,11 +6,59 @@
 
 import * as path from 'path';
 import * as fs from 'fs';
+import { window, workspace, ExtensionContext, commands, Uri, EndOfLine } from 'vscode';
+import { LanguageClient, LanguageClientOptions, ServerOptions, TransportKind } from 'vscode-languageclient';
 
-import { window, workspace, Disposable, ExtensionContext, commands, Uri } from 'vscode';
-import * as vscode from 'vscode';
+import {showUntitledWindow} from './showUntitledWindow';
+import {showMessage} from './showMessage';
+import * as Commands from './commands';
 
-import { LanguageClient, LanguageClientOptions, SettingMonitor, ServerOptions, TransportKind } from 'vscode-languageclient';
+function registerCommands(client: LanguageClient, context: ExtensionContext) {
+  context.subscriptions.push(
+    commands.registerTextEditorCommand('apiElements.parserOutput', Commands.parseOutput.bind(this, context, client))
+  );
+
+  context.subscriptions.push(
+    commands.registerCommand('apiElements.apiary.fetchApi', Commands.fetchApi.bind(this, context))
+  );
+
+  context.subscriptions.push(
+    commands.registerCommand('apiElements.apiary.logout', Commands.logout.bind(this, context))
+  );
+
+  context.subscriptions.push(
+    commands.registerTextEditorCommand('apiElements.apiary.publishApi', Commands.publishApi.bind(this, context))
+  );
+
+}
+
+function registerNotifications(client: LanguageClient) {
+  client.onNotification({ method: "openUrl" }, url =>
+    commands.executeCommand("vscode.open", Uri.parse(<string>url))
+  );
+}
+
+function registerWindowEvents() {
+  window.onDidChangeActiveTextEditor(textEditor => {
+
+    if (textEditor.document.languageId === 'API Blueprint') {
+
+      const adjustEditor = workspace.getConfiguration('apiElements').get('editor.adjustOptions');
+
+      if (adjustEditor === true) {
+        textEditor.options = {
+          insertSpaces: false,
+          tabSize: 4,
+        };
+
+        textEditor.edit(editBuilder =>
+          editBuilder.setEndOfLine(EndOfLine.LF)
+        );
+      }
+    }
+  })
+
+}
 
 export function activate(context: ExtensionContext) {
   const serverModule = context.asAbsolutePath(path.join('server', 'server.js'));
@@ -30,55 +78,10 @@ export function activate(context: ExtensionContext) {
 
   const client = new LanguageClient('Api Elements', serverOptions, clientOptions);
 
-  context.subscriptions.push(commands.registerTextEditorCommand('apiElements.parserOutput', (editor) => {
-    const statusBarDisposable = window.setStatusBarMessage("Parsing current document...");
-    client.sendRequest({ method: "parserOutput" }, editor.document.getText())
-      .then((result) => {
-        const stringifiedResult = JSON.stringify(result, null, 2);
-        const uri = Uri.parse(`untitled:${path.join(workspace.rootPath || context.extensionPath , "parseResult.json")}`);
-        return Promise.all([stringifiedResult, uri, workspace.openTextDocument(uri)]);
-      })
-      .then(([stringifiedResult, uri, textDocument]) => {
-        const edit = new vscode.WorkspaceEdit();
-        edit.insert(<Uri>uri, new vscode.Position(0, 0), <string>stringifiedResult);
-        return Promise.all([<any>textDocument, workspace.applyEdit(edit)]);
-      })
-      .then(([textDocument, editApplied]) => {
-        return window.showTextDocument(<any>textDocument, vscode.ViewColumn.One, false);
-      })
-      .then(() => {
-        statusBarDisposable.dispose();
-      })
-      .then(null, showError);
-  }));
-
-  client.onNotification({ method: "openUrl" }, (url) => {
-    commands.executeCommand("vscode.open", vscode.Uri.parse(<string>url));
-  });
-
-  window.onDidChangeActiveTextEditor((textEditor) => {
-
-    if (textEditor.document.languageId === 'API Blueprint') {
-
-      const adjustEditor = workspace.getConfiguration('apiElements').get('editor.adjustOptions');
-
-      if (adjustEditor === true) {
-        textEditor.options = {
-          insertSpaces: false,
-          tabSize: 4,
-        };
-
-        textEditor.edit((editBuilder) => {
-          editBuilder.setEndOfLine(vscode.EndOfLine.LF);
-        });
-
-      }
-    }
-  })
+  registerCommands(client, context);
+  registerNotifications(client);
+  registerWindowEvents();
 
   context.subscriptions.push(client.start());
 
-  function showError(err) {
-    window.showErrorMessage(err.message || err);
-  }
 }